Let slide trivial MP3 conversion error
Occasionally files, such as 04-25-1975-01_Track_01, for whatever reason
cause lame to be done a little before the end of the file. This causes
the conversion process to fail, as the pipe tries to push data into stdin,
and stdin doesn't want it anymore. I've changed things a bit to let this slide,
since in practice at most a fraction of a second is lost. Errors are just
as stringent as before for WAVs and FLACs.